Added dimensions to the IP6 Detector in LayOut
I created 4 section cuts on the IP6 detector in the SketchUp file. These views of the detector were then sent to LayOut where I added the dimensions of important line segments and angle measures. The PowerPoint entitled "SketchUp Cross-Sections" details the steps I took to create the section cuts in SketchUp and add dimensions in LayOuy.